Add health bar

+++ Assets/Scripts/IBreakable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IBreakable
 {
-    void ReceiveDamage(int i_ReceivedDamage);
+    void ReceiveDamage(float i_ReceivedDamage);
 }
